requirements.txt
numpy>=1.24
Pillow>=9.5
matplotlib>=3.7
tqdm>=4.66
# Optional (only for OpenAI-assisted explanations)
openai>=1.35
Flask>=3.0
